platform/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==23.0.0
celery==5.4.0
redis==5.1.1
sqlalchemy==2.0.35
psycopg[binary]==3.2.2
python-multipart==0.0.12
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
email-validator==2.2.0
